add links to docs and method for FromSql and ExecuteSqlCommand
With the advent of hyperlinks, we can now link to the primary reference for content. With that in mind, I would like to see all Microsoft document that discusses a term reference back to the primary user docs, developer user, and source code. This way the reading and fully understand the technology.